requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
matplotlib
shap
torch
transformers
pillow
scikit-fuzzy
firebase-admin
gunicorn